+++ src/main/java/com/ozii/klinika/config/SecurityConfig.java
@@ -37,8 +37,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 		http.authorizeRequests()
 		.antMatchers("/").hasRole("PATIENT")
 		.antMatchers("/doctor/**").hasRole("DOCTOR")
-		.antMatchers("/moderator/**").hasRole("MODERATOR")
-		.antMatchers("/admin/**").hasRole("ADMIN")
+		.antMatchers("/admin/**").hasAnyRole("ADMIN", "MODERATOR")
 		.and()
 		.formLogin()
 		.loginPage("/show-login-page")
