--- src/main/webapp/WEB-INF/view/moderator-med-exam.jsp
@@ -1,20 +1,0 @@
-<html>
-<head>
-<title>Ozii Clinic for MODERATOR</title>
-<!-- Reference Bootstrap files -->
-<link rel="stylesheet"
-	href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-
-</head>
-<body>
-	<h2>Moderator Zone</h2>
-	<hr>
-	<p>
-		There you will be able to edit patient's medical examinations if
-		doctor asked. <br> Coming soon... :)
-	</p>
-	<hr>
-	<a href="${pageContext.request.contextPath}/" class="btn btn-primary"
-		role="button" aria-pressed="true">Back</a>
-</body>
-</html>
